+++ src/common/Inventory.h
@@ -10,13 +10,13 @@
 #include "items/Item.h"
 
 struct ConsumableEntry {
-  uint itemId;
+  uint8_t itemId;
   Item item;
-  uint quantity;
+  uint8_t quantity;
 };
 
 struct EquipableEntry {
-  uint itemId;
+  uint8_t itemId;
   Item item;
   bool equipped;
 };
@@ -28,12 +28,12 @@ class Inventory {
 public:
   Inventory();
 
-  void addItem(Item item, uint quantity = 1u);
-  bool dropItem(uint itemId, uint quantity = 1u);
+  void addItem(Item item, uint8_t quantity = 1u);
+  bool dropItem(uint8_t itemId, uint8_t quantity = 1u);
 
-  Item &getItem(uint itemId);
-  uint totalItems();
-  bool useItem(uint itemId);
+  Item &getItem(uint8_t itemId);
+  uint8_t totalItems();
+  bool useItem(uint8_t itemId);
 
   friend std::ostream &operator<<(std::ostream &os, const Inventory &inventory);
 
@@ -43,14 +43,14 @@ public:
    * @param destination inventory
    * @param itemId for item to be displaced
    */
-  friend void exchangeItem(Inventory &origin, Inventory &destination, uint itemId, uint quantity);
+  friend void exchangeItem(Inventory &origin, Inventory &destination, uint8_t itemId, uint8_t quantity);
 
 private:
-  bool consumeItem(uint itemId);
-  bool equipItem(uint itemId);
-  bool unequipItem(uint itemId);
-  bool toggleEquip(uint itemId);
+  bool consumeItem(uint8_t itemId);
+  bool equipItem(uint8_t itemId);
+  bool unequipItem(uint8_t itemId);
+  bool toggleEquip(uint8_t itemId);
   std::vector<ConsumableEntry> consumables;
   std::vector<EquipableEntry> equipables;
-  uint lastItemId;
+  uint8_t lastItemId;
 };
